rename json to json_gen
