+++ outlines/generate/__init__.py
@@ -3,6 +3,6 @@ from .cfg import cfg
 from .choice import choice
 from .format import format
 from .fsm import fsm
-from .json import json
+from .json_gen import json
 from .regex import regex
 from .text import text
